api/requirements.txt
Flask==2.3.3
tensorflow==2.15.0
keras==2.15.0
efficientnet==1.1.1
numpy==1.26.0
Pillow==10.1.0
